add: adicionando a classe ExceptionManager. Finalizando as classes service e controller. Correcao do tipo de dado do atributo price

+++ api-produtos/src/main/java/br/com/oaktecnologia/entity/dto/ProductListDto.java
@@ -2,7 +2,9 @@ package br.com.oaktecnologia.entity.dto;
 
 import br.com.oaktecnologia.entity.Product;
 
-public record ProductListDto (Long id, String name, String description, Double price){
+import java.math.BigDecimal;
+
+public record ProductListDto (Long id, String name, String description, BigDecimal price){
 
     public ProductListDto(Product product) {
         this(product.getId(), product.getName(), product.getDescription(), product.getPrice());
